рефакторинг под доменные области
