--- app/clients/__init__.py
@@ -1,8 +1,0 @@
-from app.clients.google import GoogleClient
-from app.clients.yandex import YandexClient
-
-
-__all__ = [
-    'GoogleClient',
-    'YandexClient',
-]
