+++ app/dependency.py
@@ -8,23 +8,23 @@ from fastapi import (
     HTTPException,
 )
 
-from app.clients import GoogleClient, YandexClient
+from app.users.auth.clients import GoogleClient, YandexClient
 from app.exception import (
     TokenExpired,
     TokenNotCorrect,
 )
 from app.infrastructure.cache import get_redis_connection
 from app.infrastructure.database import get_repository
-from app.repository.cache_tasks import TaskCache
-from app.repository import (
+from app.tasks.repository.cache_tasks import TaskCache
+from app.tasks.repository import (
     TaskRepository,
-    UserRepository,
 )
-from app.service import (
+from app.tasks.service import (
     TaskService,
-    UserService,
-    AuthService,
 )
+from app.users.user_profile.repository import UserRepository
+from app.users.user_profile.service import UserService
+from app.users.auth.service import AuthService
 
 __all__ = [
     "get_repository",
