+++ app/main.py
@@ -1,6 +1,12 @@
 from fastapi import FastAPI
-from app.handlers import routers
+from app.tasks.handlers import router as tasks_router
+from app.users.auth.handlers import router as auth_router
+from app.users.user_profile.handlers import router as user_profile_router
 
 app = FastAPI()
-for router in routers:
+for router in [
+    tasks_router,
+    auth_router,
+    user_profile_router,
+]:
     app.include_router(router)
