--- app/models/__init__.py
@@ -1,13 +1,0 @@
-from app.models.task import (
-    Task,
-    Categories,
-)
-from app.models.user import (
-    UserProfile,
-)
-
-__all__ = [
-    'Task',
-    'Categories',
-    'UserProfile',
-]
