--- app/repository/__init__.py
@@ -1,9 +1,0 @@
-from app.repository.task import TaskRepository
-from app.repository.cache_tasks import TaskCache
-from app.repository.user import UserRepository
-
-__all__ = [
-    'TaskRepository',
-    'TaskCache',
-    'UserRepository',
-]
