--- app/schemas/__init__.py
@@ -1,20 +1,0 @@
-from app.schemas.task import (
-    CreateOrUpdateTaskSchema,
-    TaskSchema,
-    TasksSchema,
-)
-from app.schemas.user import (
-    UserLoginSchema,
-    UserCreateSchema,
-)
-from app.schemas.auth import GoogleUserData, YandexUserData
-
-__all__ = [
-    'CreateOrUpdateTaskSchema',
-    'TaskSchema',
-    'TasksSchema',
-    'UserLoginSchema',
-    'UserCreateSchema',
-    'GoogleUserData',
-    'YandexUserData',
-]
