--- app/service/__init__.py
@@ -1,9 +1,0 @@
-from app.service.task import TaskService
-from app.service.user import UserService
-from app.service.auth import AuthService
-
-__all__ = [
-    'TaskService',
-    'UserService',
-    'AuthService'
-]
