+++ app/tasks/repository/__init__.py
@@ -1,0 +1,7 @@
+from app.tasks.repository.task import TaskRepository
+from app.tasks.repository.cache_tasks import TaskCache
+
+__all__ = [
+    'TaskRepository',
+    'TaskCache',
+]
