+++ app/users/auth/clients/__init__.py
@@ -1,0 +1,8 @@
+from app.users.auth.clients.google import GoogleClient
+from app.users.auth.clients.yandex import YandexClient
+
+
+__all__ = [
+    'GoogleClient',
+    'YandexClient',
+]
