+++ app/users/auth/schema.py
@@ -15,3 +15,8 @@ class YandexUserData(BaseModel):
     name: str = Field(None, alias='real_name')
     default_email: str
     access_token: str
+
+
+class UserLoginSchema(BaseModel):
+    user_id: int
+    access_token: str
