+++ app/users/auth/service.py
@@ -3,18 +3,20 @@ from datetime import timedelta
 from dataclasses import dataclass
 from jose import jwt
 
-from app.clients import GoogleClient, YandexClient
+from app.users.auth.clients import GoogleClient, YandexClient
 from app.exception import (
     UserNotFoundException,
     UserNotCorrectPasswordException,
     TokenExpired,
     TokenNotCorrect,
 )
-from app.models import UserProfile
-from app.repository import UserRepository
-from app.schemas import (
-    UserLoginSchema,
+from app.users.user_profile.models import UserProfile
+from app.users.user_profile.repository import UserRepository
+from app.users.user_profile.schema import (
     UserCreateSchema,
+)
+from app.users.auth.schema import (
+    UserLoginSchema,
 )
 from app.settings import Settings
