+++ app/users/user_profile/schema.py
@@ -1,11 +1,6 @@
 from pydantic import BaseModel
 
 
-class UserLoginSchema(BaseModel):
-    user_id: int
-    access_token: str
-
-
 class UserCreateSchema(BaseModel):
     username: str | None = None
     password: str | None = None
