+++ app/users/user_profile/service.py
@@ -1,10 +1,9 @@
-
 from dataclasses import dataclass
 
-from app.models import UserProfile
-from app.repository import UserRepository
-from app.schemas import UserLoginSchema
-from app.service.auth import AuthService
+from app.users.user_profile.models import UserProfile
+from app.users.user_profile.repository import UserRepository
+from app.users.auth.schema import UserLoginSchema
+from app.users.auth.service import AuthService
 
 
 @dataclass
@@ -16,5 +15,3 @@ class UserService:
         user: UserProfile = await self.user_repository.create_user(user_name, password)
         access_token = await self.auth_service.generate_access_token(user_id=user.id)
         return UserLoginSchema(user_id=user.id, access_token=access_token)
-
-
